+++ src/main/java/http/common/HttpHeader.java
@@ -1,4 +1,4 @@
-package webserver;
+package http.common;
 
 import org.springframework.util.StringUtils;
