+++ src/main/java/http/request/RequestMethod.java
@@ -1,4 +1,4 @@
-package webserver;
+package http.request;
 
 public enum RequestMethod {
     GET(false), DELETE(false), POST(true), PUT(true);
