+++ src/main/java/webserver/RequestHandler.java
@@ -1,5 +1,7 @@
 package webserver;
 
+import http.request.HttpRequest;
+import http.response.HttpResponse;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import utils.FileIoUtils;
